Switch to manual generation of low-level wrapper
Because of the way Nimble is designed, hooks are not run when installing
from a remote source, such as Git or the Nimble package list. Since
nimbox/termbox.nim used to be generated using such a hook, Nimbox would
not work when installed from a remote source. This is now fixed, as
nimbox/termbox.nim is generated using the `wrap` Nimble task.