Add 256 Color Support (#5)
* Add 256 Color Support

This commit addresses issue #2. It adds 256 color support very naively.
It does not check for terminal support of 256 colors. It does not check
that the int provided is in the range 0..255. It does not have any error
checking or handling of any sort.

Fixes: #2

* Example for 256color support

This commit adds a worked example of 256 color support.

* Subrange Restriction for 256 Colors print proc.

By using a SubRange called Colors256, we can ensure that the print
procedure is never passed a color it cannot print. Further reading
on the 256 colors palette suggests that fixing names for the colors
probably will not work as each terminal defines them slightly
diffferently.